# contracts/openapi/mindspace.openapi.json
{
  "components": {
    "schemas": {
...
      "ChatAttachment": {
        "properties": {
          "attachment_id": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Attachment Id",
            "type": "string"
          },
          "content": {
            "default": "",
            "maxLength": 20000,
            "title": "Content",
            "type": "string"
          },
          "media_type": {
            "default": "text/plain",
            "maxLength": 120,
            "minLength": 1,
            "title": "Media Type",
            "type": "string"
          },
          "name": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "size": {
            "default": 0,
            "maximum": 1048576.0,
            "minimum": 0.0,
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "attachment_id",
          "name"
        ],
        "title": "ChatAttachment",
        "type": "object"
      },
...
      "ChatResponse": {
        "properties": {
          "assistant_message_id": {
            "default": "",
            "title": "Assistant Message Id",
            "type": "string"
          },
          "completed_at": {
            "title": "Completed At",
            "type": "string"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "llm_call_count": {
            "default": 0,
            "title": "Llm Call Count",
            "type": "integer"
          },
          "model": {
            "$ref": "#/components/schemas/ModelDiagnostics"
          },
          "model_usage": {
            "items": {
              "$ref": "#/components/schemas/ModelUsage"
            },
            "title": "Model Usage",
            "type": "array"
          },
          "presentation_mode": {
            "default": "dialogue",
            "enum": [
              "dialogue",
              "scene"
            ],
            "title": "Presentation Mode",
            "type": "string"
          },
          "reply": {
            "default": "",
            "title": "Reply",
            "type": "string"
          },
          "retrieval_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Retrieval Counts",
            "type": "object"
          },
          "round": {
            "title": "Round",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "success",
              "error"
            ],
            "title": "Status",
            "type": "string"
          },
          "tool_execution": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Execution"
          },
          "trace": {
            "items": {
              "type": "string"
            },
            "title": "Trace",
            "type": "array"
          },
          "writeback_applied": {
            "default": false,
            "title": "Writeback Applied",
            "type": "boolean"
          }
        },
        "required": [
          "session_id",
          "round",
          "status"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
      "ChatTurnCreateRequest": {
        "description": "Public chat input with server-resolved fields hidden from OpenAPI.\n\nThe hidden fields remain typed compatibility inputs so older clients keep\ntheir existing validation and idempotency behavior. ConversationService\nstill replaces every one of them with authoritative server state.",
        "properties": {
          "activity_session_id": {
            "default": "",
            "maxLength": 100,
            "title": "Activity Session Id",
            "type": "string"
          },
          "adult_mode": {
            "default": false,
            "title": "Adult Mode",
            "type": "boolean"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/ChatAttachment"
            },
            "maxItems": 5,
            "title": "Attachments",
            "type": "array"
          },
          "character_id": {
            "default": "",
            "maxLength": 64,
            "title": "Character Id",
            "type": "string"
          },
          "character_name": {
            "default": "AI助手",
            "title": "Character Name",
            "type": "string"
          },
          "client_sent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Sent At"
          },
          "client_timezone": {
            "default": "UTC",
            "title": "Client Timezone",
            "type": "string"
          },
          "client_utc_offset_minutes": {
            "default": 0,
            "maximum": 840.0,
            "minimum": -840.0,
            "title": "Client Utc Offset Minutes",
            "type": "integer"
          },
          "initiative": {
            "default": false,
            "title": "Initiative",
            "type": "boolean"
          },
          "initiative_sequence": {
            "default": 0,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Initiative Sequence",
            "type": "integer"
          },
          "initiative_sequence_limit": {
            "default": 0,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Initiative Sequence Limit",
            "type": "integer"
          },
          "initiative_trigger": {
            "default": "none",
            "enum": [
              "none",
              "manual",
              "idle_continuation",
              "continuous_companionship"
            ],
            "title": "Initiative Trigger",
            "type": "string"
          },
          "input_evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InputEvidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "interaction_mode": {
            "default": "text",
            "enum": [
              "text",
              "voice"
            ],
            "title": "Interaction Mode",
            "type": "string"
          },
          "interactions": {
            "items": {
              "$ref": "#/components/schemas/ChatInteraction"
            },
            "maxItems": 12,
            "title": "Interactions",
            "type": "array"
          },
          "message": {
            "default": "",
            "maxLength": 10000,
            "title": "Message",
            "type": "string"
          },
          "mode": {
            "default": "primary",
            "enum": [
              "primary",
              "regenerate"
            ],
            "title": "Mode",
            "type": "string"
          },
          "presentation_mode": {
            "default": "auto",
            "enum": [
              "auto",
              "dialogue",
              "scene"
            ],
            "title": "Presentation Mode",
            "type": "string"
          },
          "r18_style_id": {
            "default": "high_intensity",
            "maxLength": 64,
            "minLength": 1,
            "title": "R18 Style Id",
            "type": "string"
          },
          "reply_length_preference": {
            "default": "",
            "maxLength": 300,
            "title": "Reply Length Preference",
            "type": "string"
          },
          "reply_to_message_id": {
            "default": "",
            "maxLength": 64,
            "title": "Reply To Message Id",
            "type": "string"
          },
          "retrieval": {
            "$ref": "#/components/schemas/RetrievalSettings"
          },
          "round": {
            "default": 1,
            "minimum": 1.0,
            "title": "Round",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "user_name": {
            "default": "用户",
            "title": "User Name",
            "type": "string"
          },
          "user_persona": {
            "default": "",
            "title": "User Persona",
            "type": "string"
          },
          "voice_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VoiceInteractionContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "voice_delivery": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VoiceDeliveryState"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ChatTurnCreateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InterruptRequest": {
        "properties": {
          "request_id": {
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "InterruptRequest",
        "type": "object"
      },
...
      "SessionCreateRequest": {
        "properties": {
          "character_id": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Character Id",
            "type": "string"
          },
          "session_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "character_id"
        ],
        "title": "SessionCreateRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Mindspace Graph",
    "version": "0.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/chat": {
      "post": {
        "operationId": "chat_api_v1_chat_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatTurnCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat"
      }
    },
...
    "/api/v1/chat/stream": {
      "post": {
        "operationId": "chat_stream_api_v1_chat_stream_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-request-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatTurnCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Stream"
      }
    },
...
    "/api/v1/interrupt": {
      "post": {
        "operationId": "interrupt_api_v1_interrupt_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterruptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Interrupt"
      }
    },
...
    "/api/v1/runs/{run_id}": {
      "get": {
        "operationId": "get_run_status_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Status"
      }
    },
    "/api/v1/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_run_api_v1_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run"
      }
    },
    "/api/v1/runs/{run_id}/prompt-inspection": {
      "get": {
        "operationId": "prompt_inspection_api_v1_runs__run_id__prompt_inspection_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "reveal",
            "required": false,
            "schema": {
              "default": false,
              "title": "Reveal",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prompt Inspection"
      }
    },
    "/api/v1/runs/{run_id}/stream": {
      "get": {
        "operationId": "resume_chat_stream_api_v1_runs__run_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "last-event-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Chat Stream"
      }
    },
...
    "/api/v1/sessions": {
      "get": {
        "operationId": "list_sessions_api_v1_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Sessions"
      },
      "post": {
        "operationId": "create_session_api_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session"
      }
    },
    "/api/v1/sessions/{session_id}": {
      "delete": {
        "operationId": "delete_session_api_v1_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Session"
      },
      "get": {
        "operationId": "get_session_api_v1_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session"
      }
    },
    "/api/v1/sessions/{session_id}/clear": {
      "post": {
        "operationId": "clear_session_api_v1_sessions__session_id__clear_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Session"
      }
    },
    "/api/v1/sessions/{session_id}/context-diagnostics": {
      "get": {
        "operationId": "context_diagnostics_api_v1_sessions__session_id__context_diagnostics_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Context Diagnostics"
      }
    },
    "/api/v1/sessions/{session_id}/messages/{message_id}": {
      "delete": {
        "operationId": "delete_message_api_v1_sessions__session_id__messages__message_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Message"
      }
    },
    "/api/v1/sessions/{session_id}/rounds/{round_num}": {
      "delete": {
        "operationId": "delete_round_api_v1_sessions__session_id__rounds__round_num__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "round_num",
            "required": true,
            "schema": {
              "title": "Round Num",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Round"
      }
    },
...
  }
}